set logger level via flag, refactor
